# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Paperless AI API",
    "version": "1.11.0",
    "description": "API for Paperless AI document processing"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Server"
    }
  ],
  "components": {
    "schemas": {
...
      "AiModelListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "modelIdentifier": {
            "type": "string"
          },
          "inputTokenPrice": {
            "type": ["number", "null"]
          },
          "outputTokenPrice": {
            "type": ["number", "null"]
          },
          "isActive": {
            "type": "boolean"
          },
          "aiAccountId": {
            "type": "string"
          },
          "aiAccount": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "provider": {
                "type": "string"
              }
            },
            "required": ["id", "name", "provider"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "modelIdentifier",
          "inputTokenPrice",
          "outputTokenPrice",
          "isActive",
          "aiAccountId",
          "aiAccount",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateAiModelRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "modelIdentifier": {
            "type": "string",
            "minLength": 1
          },
          "aiAccountId": {
            "type": "string",
            "minLength": 1
          },
          "inputTokenPrice": {
            "type": ["number", "null"],
            "exclusiveMinimum": 0
          },
          "outputTokenPrice": {
            "type": ["number", "null"],
            "exclusiveMinimum": 0
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "params": {
            "type": "object",
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "string",
                  "format": "date"
                }
              ]
            }
          }
        },
        "required": ["error", "message"]
      },
...
    },
...
  },
  "paths": {
...
    "/ai-models/{id}": {
      "get": {
        "summary": "Get AiModel by ID (Admin only)",
        "tags": ["AiModels"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "AiModel details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiModelListItem"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update AiModel (Admin only)",
        "tags": ["AiModels"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAiModelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AiModel updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiModelListItem"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete AiModel (Admin only)",
        "tags": ["AiModels"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "AiModel deleted"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/ai-models/{id}/sharing/{accessId}": {
      "delete": {
        "summary": "Remove share from AiModels",
        "tags": ["AiModels"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "accessId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Share removed"
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}